Move getUrlParams to /logic
